@@ -1,71 +1,54 @@
 source 'https://rubygems.org'
 source 'https://gems.gemfury.com/vo6ZrmjBQu5szyywDszE/'
 
-if ENV['BUNDLE_DEV']
-  gem 'gds-sso', path: '../gds-sso'
-else
-  gem 'gds-sso', '0.7.9'
-end
-
-gem 'gds-warmup-controller'
-
-group :passenger_compatibility do
-  gem 'rack', '1.3.5'
-  gem 'rake', '0.9.2'
-end
-
-gem 'rails', '3.1.3'
 gem 'aws-ses', require: 'aws/ses'
-
-gem 'erubis'
-gem 'plek', '~> 0'
-gem 'gelf'
-gem 'graylog2_exceptions'
-gem 'rest-client'
-
-gem 'null_logger'
-gem 'rummageable'
-gem 'daemonette', git: 'git@github.com:alphagov/daemonette.git'
-gem 'stomp', '1.1.9'
-if ENV['MARPLES_DEV']
-  gem 'marples', path: '../marples'
-else
-  gem 'marples', '~> 1'
-end
-gem 'router-client', '2.0.3', require: 'router/client'
-
-gem 'rest-client'
-gem "colorize", "~> 0.5.8"
-
-gem 'inherited_resources'
-gem 'formtastic', git: 'git://github.com/justinfrench/formtastic.git', branch: '2.1-stable'
-gem 'formtastic-bootstrap', git: 'git://github.com/cgunther/formtastic-bootstrap.git', branch: 'bootstrap-2'
-gem 'has_scope'
-gem 'kaminari', '0.13.0'
-
-if ENV['CONTENT_MODELS_DEV']
-  gem "govuk_content_models", :path => '../govuk_content_models'
-else
-  gem "govuk_content_models", "0.1.5"
-end
-
 if ENV['CDN_DEV']
   gem 'cdn_helpers', path: '../cdn_helpers'
 else
   gem 'cdn_helpers', '0.9'
 end
+gem 'colorize', "~> 0.5.8"
+if ENV['BUNDLE_DEV']
+  gem 'gds-sso', path: '../gds-sso'
+else
+  gem 'gds-sso', '~> 1.2.0'
+end
+if ENV['CONTENT_MODELS_DEV']
+  gem "govuk_content_models", :path => '../govuk_content_models'
+else
+  gem "govuk_content_models", "1.6.4"
+end
 
+gem 'erubis'
+gem 'exception_notification', '~> 2.4.1', require: 'exception_notifier'
+gem 'formtastic', git: 'https://github.com/justinfrench/formtastic.git', branch: '2.1-stable'
+gem 'formtastic-bootstrap', git: 'https://github.com/cgunther/formtastic-bootstrap.git', branch: 'bootstrap-2'
+gem 'gds-api-adapters', '0.2.3'
+gem 'gds-warmup-controller'
+gem 'gelf'
 if ENV['GOVSPEAK_DEV']
   gem 'govspeak', path: '../govspeak'
 else
-  gem 'govspeak', '~> 0.8.15'
+  gem 'govspeak', '1.0.0'
 end
-
-gem 'exception_notification', '~> 2.4.1', require: 'exception_notifier'
-
+gem 'graylog2_exceptions'
+gem 'has_scope'
+gem 'inherited_resources'
+gem 'kaminari', '0.13.0'
 gem 'lockfile'
-gem 'whenever'
+gem 'lograge'
+gem 'mongo', '1.6.2'  # Locking this down to avoid a replica set bug
 gem 'newrelic_rpm'
+gem 'null_logger'
+gem 'plek', '0.1.24'
+gem 'rails', '3.2.8'
+gem 'rest-client'
+gem 'rest-client'
+gem 'statsd-ruby', '1.0.0'
+gem 'whenever'
+
+gem 'jquery-rails'
+gem 'less-rails-bootstrap'
 
 group :assets do
   gem "therubyracer", "~> 0.9.4"
@@ -81,9 +64,7 @@ group :test do
   gem 'shoulda'
   gem 'database_cleaner'
 
-  gem 'cucumber-rails', require: false
-
-  gem 'capybara', '~> 1.0.0'
+  gem 'capybara', '~> 1.1.0'
   gem "capybara-webkit"
   gem 'launchy'
 
@@ -94,7 +75,7 @@ group :test do
 
   gem "timecop"
 
-  gem 'simplecov', '0.4.2'
-  gem 'simplecov-rcov'
+  gem 'simplecov', '~> 0.6.4', :require => false
+  gem 'simplecov-rcov', '~> 0.2.3', :require => false
   gem 'ci_reporter'
 end
