@@ -1,3 +1,5 @@
+require "statsd"
+
 class Admin::EditionsController < Admin::BaseController
   actions :create, :update, :destroy
   defaults :resource_class => Edition, :collection_name => 'editions', :instance_name => 'resource'
@@ -15,6 +17,7 @@ class Admin::EditionsController < Admin::BaseController
 
   def create
     class_identifier = params[:edition].delete(:kind).to_sym
+    Statsd.new(::STATSD_HOST).increment("publisher.edition.create.#{class_identifier}")
     @publication = current_user.create_edition(class_identifier, params[:edition])
 
     if @publication.persisted?
@@ -28,7 +31,7 @@ class Admin::EditionsController < Admin::BaseController
   end
 
   def duplicate
-    new_edition = current_user.new_version(resource)
+    new_edition = current_user.new_version(resource, (params[:to] || nil))
     assigned_to_id = (params[:edition] || {}).delete(:assigned_to_id)
     if new_edition and new_edition.save
       update_assignment new_edition, assigned_to_id
@@ -81,7 +84,13 @@ class Admin::EditionsController < Admin::BaseController
         return redirect_to admin_edition_path(resource), :alert => "Couldn't #{params[:activity].to_s.humanize.downcase} for #{description(resource).downcase}. The email addresses you entered appear to be invalid."
       end
     end
+    current_status = resource.state
+    intended_status = params[:activity][:request_type]
     if current_user.progress(resource, params[:activity].dup)
+      statsd = Statsd.new(::STATSD_HOST)
+      statsd.decrement("publisher.edition.#{current_status}")
+      statsd.increment("publisher.edition.#{intended_status}")
+
       redirect_to admin_edition_path(resource), :notice => success_message(params[:activity][:request_type])
     else
       redirect_to admin_edition_path(resource), :alert => failure_message(params[:activity][:request_type])
