@@ -27,21 +27,20 @@ module Publisher
     config.autoload_paths += %W(#{config.root}/lib #{config.root}/app/presenters #{config.root}/app/decorators)
 
     config.assets.precompile += %W(
-        application.css bootstrap.min.css bootstrap-responsive.min.css
-
-        application.js bootstrap.min.js form-tools.js jquery.min.js
-        jquery-ui.custom.min.js
-        jquery.mustache.js multi-part.js publications.js
-
-        glyphicons-halflings-white.png glyphicons-halflings.png
-        header-crown.png icon-answer.png icon-guide.png icon-localtransaction.png
-        icon-place.png icon-programme.png icon-service.png icon-transaction.png)
+      form-tools.js
+      jquery-ui.custom.min.js
+      jquery.mustache.js
+      multi-part.js
+      publications.js
+    )
 
     config.generators do |g|
         g.orm :mongoid
         g.template_engine :erb # this could be :haml or whatever
         g.test_framework :test_unit, :fixture => false # this could be :rpsec or whatever
     end
+
+    config.action_dispatch.rack_cache = nil
 
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
