@@ -2,11 +2,11 @@ require 'test_helper'
 
 class GuideEditionTest < ActiveSupport::TestCase
   setup do
-    panopticon_has_metadata("id" => '1234574', "name" => "Childcare", "slug" => "childcare")
+    @artefact = FactoryGirl.create(:artefact, name: "Childcare", slug: "childcare")
   end
 
   def template_guide
-    edition = FactoryGirl.create(:guide_edition, slug: "childcare", title: "One", panopticon_id: 1234574)
+    edition = FactoryGirl.create(:guide_edition, slug: "childcare", title: "One", panopticon_id: @artefact.id)
     edition.start_work
     edition.save
     edition
@@ -16,7 +16,7 @@ class GuideEditionTest < ActiveSupport::TestCase
     user = User.create(:name => "Ben")
     other_user = User.create(:name => "James")
 
-    guide = user.create_edition(:guide, :panopticon_id => 1234574, :overview => 'My Overview', :title => 'My Title', :slug => 'my-title', :alternative_title => 'My Other Title')
+    guide = user.create_edition(:guide, :panopticon_id => FactoryGirl.create(:artefact).id, :overview => 'My Overview', :title => 'My Title', :slug => 'my-title', :alternative_title => 'My Other Title')
     edition = guide
     user.start_work(edition)
     user.request_review(edition,{:comment => "Review this guide please."})
@@ -24,6 +24,7 @@ class GuideEditionTest < ActiveSupport::TestCase
     user.send_fact_check(edition,{:comment => "Review this guide please.", :email_addresses => 'test@test.com'})
     user.receive_fact_check(edition, {:comment => "No changes needed, this is all correct"})
     other_user.approve_fact_check(edition, {:comment => "Looks good to me"})
+    stub_register_published_content
     user.publish(edition, {:comment => "PUBLISHED!"})
     return user, guide
   end
@@ -33,33 +34,6 @@ class GuideEditionTest < ActiveSupport::TestCase
     edition.parts.build
     edition.parts.build(:order => 1)
     assert edition.order_parts
-  end
-
-  test "struct for search index" do
-    edition = template_guide
-    edition.update_attribute(:state, 'published')
-    data = edition.search_index
-    assert_equal ["title", "link", "format", "description", "indexable_content", "section", "subsection", "additional_links"], data.keys
-    assert_equal edition.title, data['title']
-    assert_equal "guide", data['format']
-  end
-
-  test "indexable content contains parts for search index" do
-    edition = template_guide
-    edition.update_attribute(:state, 'published')
-    edition.parts.build(:body => "ONE", :title => "ONE", :slug => "/one")
-    edition.parts.build(:body => "TWO", :title => "TWO", :slug => "/two")
-    data = edition.search_index
-    assert_equal "ONE ONE TWO TWO", data['indexable_content']
-  end
-
-  test "index contains parts as additional links on published guide" do
-    edition = template_guide
-    edition.update_attribute(:state, 'published')
-    edition.parts.build(:body => "ONE", :title => "ONE", :slug => "one")
-    edition.parts.build(:body => "TWO", :title => "TWO", :slug => "two")
-    data = edition.search_index
-    assert_equal 2, data['additional_links'].count
   end
 
   test 'a guide without a video url should not have a video' do
