@@ -7,7 +7,7 @@ class ProgrammeEditionTest < ActiveSupport::TestCase
   end
 
   def template_programme
-    p = ProgrammeEdition.new(:slug=>"childcare", :title=>"Children", :panopticon_id => 987353)
+    p = ProgrammeEdition.new(:slug=>"childcare", :title=>"Children", :panopticon_id => FactoryGirl.create(:artefact).id)
     p.start_work
     p.save
     p
@@ -19,17 +19,6 @@ class ProgrammeEditionTest < ActiveSupport::TestCase
     g.parts.build
     g.parts.build(:order => 1)
     assert g.order_parts
-  end
-
-  test 'a programme correctly formats the additional links' do
-    programme = template_programme
-    programme.update_attribute(:state,'published')
-    programme.save
-
-    out = programme.search_index
-    assert_equal 5, out['additional_links'].count
-    assert_equal '/childcare#overview', out['additional_links'].first['link']
-    assert_equal '/childcare/further-information', out['additional_links'].last['link']
   end
 
   test "new programme has correct parts" do
