@@ -4,7 +4,7 @@ class UserTest < ActiveSupport::TestCase
   test "it doesn't try to send a fact check email if no addresses were given" do
     user = User.create(:name => "bob")
     NoisyWorkflow.expects(:request_fact_check).never
-    trans = user.create_edition(:transaction, title: "test answer", slug: "test", panopticon_id: 123)
+    trans = user.create_edition(:transaction, title: "test answer", slug: "test", panopticon_id: FactoryGirl.create(:artefact).id)
     assert ! user.send_fact_check(trans, {comment: "Hello"})
   end
 
@@ -12,12 +12,16 @@ class UserTest < ActiveSupport::TestCase
     user = User.create(:name => "bob")
     second_user = User.create(:name => "dave")
 
-    trans = user.create_edition(:transaction, title: "test answer", slug: "test", panopticon_id: 123)
+    trans = user.create_edition(:transaction, title: "test answer", slug: "test", panopticon_id: FactoryGirl.create(:artefact).id)
     user.start_work(trans)
     user.request_review(trans, {comment: "Hello"})
     second_user.approve_review(trans, {comment: "Hello"})
 
-    Messenger.instance.expects(:published).with(trans).once
+    stub_register_published_content
     user.publish trans, {comment: "Published because I did"}
+  end
+
+  test "use a custom collection for users" do
+    assert_equal "publisher_users", User.collection_name
   end
 end
